+++ src/test/main.ts
@@ -16,6 +16,7 @@ import "./regression.test.js";
 import "./integration.test.js";
 import "./ui.test.js";
 import "./round5.test.js";
+import "./round6.test.js";
 
 import { run } from "./runner.js";
 run();